Detected some electronical inteferences like coffe machines, blind motors, .. can cause false positive taking a photo. Added method to be sure that button is pressed.

+++ PhotoBooth.py
@@ -11,10 +11,11 @@ compose_and_print_process = None
 
 def cancel_print(bin_var):
     global compose_and_print_process
-
-    os.killpg(os.getpgid(compose_and_print_process.pid), signal.SIGTERM)
-    picam.cancel_countdown()
-
+    #Detected some electronical devices generated interferences causing falses positives. Due to this fact, we check that really button has been pressed
+    #before cancel print
+    if (input_cancel_print_button.check_button_is_pressed()):
+        os.killpg(os.getpgid(compose_and_print_process.pid), signal.SIGTERM)
+        picam.cancel_countdown()
 
 def composed_and_print_captured_image(photo_path):
     global compose_and_print_process
@@ -23,16 +24,21 @@ def composed_and_print_captured_image(photo_path):
     compose_and_print_process = subprocess.Popen('./ComposeAndPrint.sh ' + photo_path + ' ' + composed_photo_path, stdout=subprocess.PIPE, shell=True, preexec_fn=os.setsid)
 
 
+CAPTURE_BUTTON_CHANNEL = 4
+CANCEL_PRINT_BUTTON_CHANNEL = 11
+LED_CAPTURE_BUTTON_CHANNEL = 18
+LED_CANCEL_PRINT_BUTTON_CHANNEL = 7
+
 picam = Camera()
 
 picam.framerate = 25
 picam.start_preview()
 
-blinking_led_capture_button = BlinkingLed(18, 0.5)
-blinking_led_cancel_print_button = BlinkingLed(7, 0.5)
+blinking_led_capture_button = BlinkingLed(LED_CAPTURE_BUTTON_CHANNEL, 0.5)
+blinking_led_cancel_print_button = BlinkingLed(LED_CANCEL_PRINT_BUTTON_CHANNEL, 0.5)
 
-input_capture_button = InputButton(4)
-input_cancel_print_button = InputButton(11)
+input_capture_button = InputButton(CAPTURE_BUTTON_CHANNEL)
+input_cancel_print_button = InputButton(CANCEL_PRINT_BUTTON_CHANNEL)
 
 try:
